// src/lib/models/Organization.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Organization = {
    id?: string;
    businessActorId?: string;
    code?: string;
    name?: string;
    serviceType?: string;
    status?: string;
    managerId?: string;
    email?: string;
    description?: string;
    logoUri?: string;
    logoId?: string;
    websiteUrl?: string;
    socialNetwork?: string;
    businessRegistrationNumber?: string;
    taxNumber?: string;
    capitalShare?: number;
    ceoName?: string;
    yearFounded?: number;
    foundingDate?: string;
    legalForm?: string;
    keywords?: string;
    numberOfEmployees?: number;
    createdAt?: string;
    updatedAt?: string;
    active?: boolean;
    individualBusiness?: boolean;
};